Tweaks

+++ scripts/20_brew/macOS/sync.sh
@@ -11,7 +11,7 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 
 # Arm binaries are stored under /opt/homebrew/bin while x86 binaries are stored under /usr/local/bin.
 # On x86, create symlink /opt/homebrew/bin pointing to /usr/local/bin to allow to use path /opt/homebrew/bin all the time.
-if [ "$(uname -p)" = "i386" ]; then
+if [ ! -e /opt/homebrew ] && [ "$(uname -p)" = "i386" ]; then
   echo "Creating symlink for binary folder..."
   sudo mkdir -p /opt/homebrew
   sudo ln -s /usr/local/bin /opt/homebrew/bin
